refactor(team): extract passive user reply summary linking from the message feed service
Move the passive user reply summary linking cluster out of
TeamMessageFeedService into its own module.

New module src/main/services/team/linkPassiveUserReplySummaries.ts holds
linkPassiveUserReplySummaries plus the three pieces only it uses: the
PASSIVE_USER_REPLY_LINK_WINDOW_MS constant and the module-private helpers
normalizePassiveUserReplyLinkText and extractPassiveUserPeerSummaryBody.
Only linkPassiveUserReplySummaries is exported; the helpers stay private.

This is pure code motion. Every moved line is byte-identical to its
previous form; the single textual change is the added `export` keyword on
the linkPassiveUserReplySummaries declaration. The service keeps both call
sites verbatim (`messages = linkPassiveUserReplySummaries(messages);` in
getPage and in buildFeed) and gains one import; the now-unused
classifyIdleNotificationText import moves to the new module. No test needed
a change, which is the proof there is no behaviour change.

TeamMessageFeedService.ts goes from 1094 to 986 lines; the new module is
112 lines. No entry in scripts/ci/source-file-size-baseline.json is touched.

Note for reviewers: an identical but completely dead copy of this cluster
also lives in TeamDataService.ts (constant, both helpers, and an unused
private method). It is deliberately left alone here, since deleting dead
code is not code motion.
